### 2024-02-07 #6329 @dfm
In the meantime, I believe that the errors raised here are happening because the paper must be called `paper.md` [as described in the docs](https://joss.readthedocs.io/en/latest/submitting.html). Please rename the file and try commenting `@editorialbot generate pdf` in this thread.
