### 2024-03-13 #6329 @warrickball
Hi @darthoctopus, sorry about the long wait. I'll pick this up now and start looking for reviewers. If there's anyone you'd like to suggest that don't constitute a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), do let me know!
