### 2024-03-25 #6329 @warrickball
I just want to flag for @dfm that I'm technically a conflict of interest myself. @darthoctopus and I have both been authors on some consortium papers organised by the TESS Asteroseismic Science Consortium (TASC) but we've also both contributed to [PBjam](https://github.com/grd349/PBjam/), on which Reggae depends. My contributions to PBjam have been very limited, though.

Regarding the code, Reggae doesn't clearly come across clearly as a tool intended for other researchers using PBjam. You may want to have a look at the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) in anticipation. For example, I couldn't easily find [installation instructions](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions) or a worked example. The GUI screenshot in the `README.md` is fine but it'd be nice to know how to reach that point.
