### 2024-03-26 #6329 @warrickball
You can certainly iterate during review and I think most authors do respond to the review by making changes to the repository. The JOSS review process is intended to be very iterative and reviewers may comment on one aspect of the code at a time, so that you're working on one part (e.g. documentation) while they review another (e.g. testing).
