### 2024-04-08 #6588 @editorialbot
[REVIEW]: Reggae: A Parametric Tuner for PBJam, and a Visualization Tool for Red Giant Oscillation Spectra
**Submitting author:** @darthoctopus (<a href="http://orcid.org/0000-0001-7664-648X">Joel Ong J. M.</a>)
**Repository:** https://github.com/darthoctopus/reggae
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.0
