### 2024-04-08 #6588 @sybreton
Hi @dfm,

Before we go on, can you formally confirm that the COIs I might have with @darthoctopus have been waived ?
We have been co-authors on the following large collaboration papers:
https://ui.adsabs.harvard.edu/abs/2023AJ....166..167M/abstract
https://ui.adsabs.harvard.edu/abs/2023AJ....165..214C/abstract
https://ui.adsabs.harvard.edu/abs/2022AJ....163...79H/abstract
https://ui.adsabs.harvard.edu/abs/2021AJ....162..211H/abstract
but when he contacted me @warrickball told me that this should not represent an issue.
