### 2024-04-08 #6588 @dfm
@sybreton — Many thanks for bringing this up! Like @warrickball says, this type of co-authorship within a large collaboration can typically be waived unless anyone involved in the review has any concerns (please feel free to send me an email or comment here if you have any!). The only comment I would have is that the COI policy should apply to all co-authors of this submission, so if you anticipate any larger conflicts there, please let me know! Thanks again!!
