### 2024-04-09 #6588 @sybreton
Thanks for the information !
I went to check the co-author list in order to check possible COI, I hope everything should be fine:
Martin Nielsen is also among the co-authors of one of the paper already mentioned above:
https://ui.adsabs.harvard.edu/abs/2023AJ....166..167M/abstract
Martin Nielsen, Guy Davies and I are members of the PLATO consortium:
https://platomission.com/the-plato-consortium/plato-consortium-members/
Nevertheless, we are not involved on the same PLATO working packages, although I think I should mention that Martin is the author of a few functions that are included in my open source project related to the mission preparation:
https://gitlab.com/sybreton/star_privateer/-/blob/main/src/star_privateer/lomb_scargle.py?ref_type=heads
I have no perceived COI with Emily Hatt.
