### 2024-04-10 darthoctopus/reggae#9 @sybreton
Need to expand the documentation to meet JOSS review criteria
Hi @darthoctopus et al.,
I have started the JOSS review and I found that you will need to significantly expand the current documentation you provide with the module so the review "Documentation" criteria can be met.

In particular:
- in the current README, you directly mention the implementation you provide with the module, can you add beforehand a statement of need similar to the one you provide for the software paper ?
- can you add a new section describing the installation instructions to follow for the module ?
- can you provide an example usage of the module (for example through a "Quickstart" tutorial) ? As far as I understand, the module functionality are accessible through a GUI, a description of the possibilities offered by this interface would be very useful.
- can you add an API description of the main functions implemented in the module ?
- I see that the repository already has a `test` repository, can you explain how to run these tests, and, if necessary, expand the scope of the tests so the coverage of the source code to be tested is as expanded as possible ?
- can you add the community guidelines in the documentation ?

You can take a look at the checklist below to see the elements that need to be added to the documentation in order to comply with the review criteria:
https://github.com/openjournals/joss-reviews/issues/6588#issuecomment-2042632815

If you need any example in order to expand the documentation in the direction stated above, I also provide a link to the `corner` module documentation, which provide a good and compact example of a documentation with installation and test instruction, quickstart tutorials and API description.
https://corner.readthedocs.io/en/latest/

Please do not hesitate to reach out to me if you need some clarifications or have additional questions,

openjournals/joss-reviews#6588
