+++ reggae/utils.py
@@ -4,19 +4,25 @@ import jax.numpy as jnp
 import jax.scipy.special as jsp
 import jax
 import numpy as np
-import lightkurve as lk
 import os
 jax.config.update('jax_enable_x64', True)
 
 
 def readBkgFits(ID):
-    """ Read background samples for bog prior
+    """ Read background samples during prior.
     
     Read in the previosly fit background model parameters to use as a prior
-    on the bog sampling. 
+    during sampling. 
 
-    This will be removed once the background parameters are put into the prior
-    and become part of the DR.
+    Paramaters
+    ----------
+    ID : str
+        Target identifier
+
+    Returns
+    -------
+    out : dict
+        Dictionary containing parameters for three harvey laws.
     """  
     
     fname = os.path.join(*['Examples', ID+'_bkg_samples.npz'])
@@ -43,10 +49,12 @@ def readBkgFits(ID):
     return out
 
 
-
-# Jaxxed
 class jaxInterp1D():
-    """ Replacement for scipy.interpolate.interp1d in jax"""
+    """ Replacement for scipy.interpolate.interp1d in jax
+
+    This is used to initialize a callable object for interpolating.
+    
+    """
 
     def __init__(self, xp, fp, left=None, right=None, period=None):
         self.xp = xp
@@ -57,7 +65,19 @@ class jaxInterp1D():
 
     @partial(jax.jit, static_argnums=(0,))
     def __call__(self, x):
+        """ Evalute the interpolant on given values
 
+        Parameters
+        ----------
+        x : jnp.array
+            jax array of points to evaluate the interpolant at.
+        
+        Returns
+        -------
+        interp : jnp.array
+            The interpolated values.
+
+        """
         return jnp.interp(x, self.xp, self.fp, self.left, self.right, self.period)
 
 
@@ -91,7 +111,7 @@ class scalingRelations():
         Parameters
         ----------
         nu : float, array
-            Value(s) at which to compute dnu, assuming nu corresponds to numax.
+            Value(s) at which to compute dnu, assuming nu corresponds to numax. (muHz)
         gamma : float
             Scaling factor to apply.
         p : array-like, optional
@@ -110,30 +130,23 @@ class scalingRelations():
     def envWidth(self, numax):
         """ Scaling relation for the envelope width
 
-        Computest he full width at half maximum of the p-mode envelope based
-        on numax and Teff (optional).
+        Computes the the estimated full width at half maximum of the p-mode 
+        envelope based on numax.
 
         Parameters
         ----------
         numax : float
             Frequency of maximum power of the p-mode envelope.
-        Teff : float, optional
-            Effective surface temperature of the star.
-        Teff0 : float, optional
-            Solar effective temperature in K. Default is 5777 K.
-
+        
         Returns
         -------
         width : float
             Envelope width in muHz
         """
 
-        width = 0.66*numax**0.88 # Mosser et al. 201??
+        width = 0.66*numax**0.88
 
         return width
-
-# Jaxxed
-
 
 @jax.jit
 def to_log10(x, xerr):
@@ -156,8 +169,6 @@ def to_log10(x, xerr):
 
     """
 
-
-
     L = jax.lax.cond(xerr > 0,
                      lambda A: jnp.array([jnp.log10(A[0]), A[1]/A[0]/jnp.log(10.0)]),
                      lambda A: A,
@@ -165,14 +176,16 @@ def to_log10(x, xerr):
 
     return L
 
-# Jaxxed
 class normal():
 
     def __init__(self, **kwargs):
         """ normal distribution class
 
-        Create instances a probability density which follows the normal
+        Create instances of a probability density which follows a normal
         distribution.
+
[…]
