+++ reggae/theta.py
@@ -1,25 +1,26 @@
 from .utils import beta, normal
 from dataclasses import dataclass, astuple
 import jax.numpy as jnp
+import scipy
 
 @dataclass
 class ThetaReg:
     '''
     Parameters for Reggae (asymptotic description of l = 1 modes)
 
-    If θ is a numpy array, generate an instance as
+    If theta is a numpy array, generate an instance as
     theta = ThetaReg(*θ)
     '''
 
-    dPi0: float
-    p_L: float
-    p_D: float
-    epsilon_g: float
-    log_omega_core: float
-    d01: float
-    alpha_g: float
-    inclination: float
-    log_omega_env: float
+    dPi0: float # Period spacing
+    p_L: float # Coupling parameter
+    p_D: float # Coupling parameter
+    epsilon_g: float # g-mode offset
+    log_omega_core: float # log10 core splitting
+    d01: float # l=0,1 frequency splitting
+    alpha_g: float # g-mode curvature term
+    inclination: float # inclination stellar rotation axis
+    log_omega_env: float # log10 envelope splitting
 
     dims = 9
 
@@ -41,9 +42,24 @@ class ThetaReg:
     beta12 = beta(a=1, b=2)
     normal = normal(mu=0, sigma=1)
 
-    #TODO: jax this up
+    
     @staticmethod
     def prior_transform(u, bounds=None):
+        """ Prior transform for inverse sampling
+
+        Evalues the ppf (quantile function) given a set of quantile values u
+        drawn from the n-dimensional hypercube. 
+
+        Parameters
+        ----------
+        u : np.array
+            Array of values between 0 and 1 drawn uniformly from the 
+            n-dimensional hypercube.
+        bounds : np.array
+            Array of bounds for the distributions in case they need to be
+            truncated.
+        """
+
         if bounds is None:
             bounds = ThetaReg.bounds
         θ = [a + (b-a)*t for (a, b), t in zip(bounds, u)]
@@ -56,6 +72,21 @@ class ThetaReg:
 
     @staticmethod
     def inv_prior_transform(θ, bounds=None):
+        """ Prior transform for inverse sampling
+
+        Evalues the ppf (quantile function) given a set of quantile values u
+        drawn from the n-dimensional hypercube. 
+
+        Parameters
+        ----------
+        u : np.array
+            Array of values between 0 and 1 drawn uniformly from the 
+            n-dimensional hypercube.
+        bounds : np.array
+            Array of bounds for the distributions in case they need to be
+            truncated.
+        """
+        
         if bounds is None:
             bounds = ThetaReg.bounds
         x = jnp.array([(ξ - a) / (b - a) for (a, b), ξ in zip(bounds, θ)])
@@ -68,30 +99,37 @@ class ThetaReg:
 
 @dataclass
 class ThetaAsy:
-    log_numax: float
-    log_dnu: float
-    eps: float
-    log_d02: float
-    log_alpha: float
-    log_hmax: float
-    log_env_width: float
-    log_mode_width: float
+    """ Parameters for the asymptotic description of l = 0,2 modes
+    """
+
+    log_numax: float # log10 numax
+    log_dnu: float # log10 dnu
+    eps: float # p-mode phase offset
+    log_d02: float # log10  l=0,2 frequency spacing
+    log_alpha: float # log10 p-mode curvature
+    log_hmax: float # log10 envelope height
+    log_env_width: float # log10 envelope width
+    log_mode_width: float # log10 mode width
 
     dims = 8
 
     def nmax(self):
+        """Compute nmax"""
+
         numax = 10.**self.log_numax
         dnu = 10.**self.log_dnu
         eps = self.eps
[…]
