+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. reggae documentation master file, created by
+   sphinx-quickstart on Wed May  1 06:14:41 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to reggae's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
