+++ docs/index.rst
@@ -7,10 +7,12 @@ Welcome to reggae's documentation!
 ==================================
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
    :caption: Contents:
 
-
+   installation
+   usage
+   api
 
 Indices and tables
 ==================
