+++ docs/usage.md
@@ -1,0 +1,3 @@
+# Usage
+
+How 3 use
