### 2024-05-01 fb0bcb9
polish docstrings
