+++ reggae/psdmodel.py
@@ -1,18 +1,40 @@
-import numpy as np
 import jax.numpy as jnp
-from .theta import ThetaReg
-
 from .reggae import reggae
 from .asymptotic import asymptotic
 
 class PSDModel(reggae, asymptotic):
 
     def __init__(self, f, n_orders, lw=1/200, nu_0=None, nu_2=None, *args, **kwargs):
+        """ Creates an instance of the spectrum model
+
+        Consists of a set of l=0 modes and a set of multiplets for l=1 and l=2 modes.
+
+        The l=1 modes are treated as mixed. The l=2 more are assumed to not be mixed.
+        
+        Parameters
+        ----------
+        f: array-like
+            Frequencies at which to compute the model spectrum.
+        n_orders
+            Number of radial p-mode orders to include in the model.
+        lw: float
+            The linewidth of the p-modes.
+        nu_0: array-like
+            The frequencies of l=0 modes, with length n_orders.
+        nu_2: array-like
+            The frequencies of l=0 modes, with length n_orders.
+        """
+
         self.f = f
+
         self.n_orders = n_orders
+        
         self.lw = lw
+        
         self.n_g = None
+        
         self.nu_0 = nu_0
+        
         self.nu_2 = nu_2
 
     # def init_psdmodel(self, f, numax, nu_0, d02, env_width, alpha_p=0, lw=1/50):
@@ -36,64 +58,168 @@ class PSDModel(reggae, asymptotic):
     #     self.n_g = None
 
     def get_nu_0(self, theta_asy):
+        """ Return the l=0 p-mode frequencies
+        
+        Parameters
+        ----------
+        theta_asy: data class
+            Data class containing the asymptotic parameters for p-modes. 
+
+        Returns
+        -------
+        nu0: array-like
+            l=0 mode frequencies.
+        """
+
         if self.nu_0 is not None:
             return self.nu_0
+        
         return theta_asy.nu_0(self.n_orders)
 
-    def get_nu_2(self, theta_asy):
+    def get_nu_2(self,):
+        """ Get the l=2 mode frequencies
+
+        Returns
+        -------
+        nu0: array-like
+            l=2 mode frequencies
+        """
+
         if self.nu_2 is not None:
             return self.nu_2
 
     def get_d02(self, theta_asy):
-        if self.nu_2 is not None and self.nu_0 is not None:
+        """ Compute the l=0,2 separation
+        
+        Parameters
+        ----------
+        theta_asy: data class
+            Data class of asymtotic p-mode parameters
+
+        Returns
+        -------
+        d02: float
+            Array of l=0 mode frequencies
+        """
+
+        if self.nu_2 is not None and self.nu_0 is not None: # TODO Aren't these checks redundant?
             return self.get_nu_0(None) - self.get_nu_2(None)
+        
         return 10.**(theta_asy.log_d02)
 
     def getl1(self, theta_asy, theta_reg, **kwargs):
+        """Get the mixed l=1 mode frequencies.
+
+        Parameters
+        ----------
+        theta_asy: data class
+            Data class containing the asymptotic p-mode parameters.
+        theta_reg: data class
+            Data class containing the g-mode and mixing parameters.
+
+        Returns
+        -------
+        nu_1: array-like
+            An array of n_g + n_p mode frequencies.
+        """
 
         numax = 10.**(theta_asy.log_numax)
+
         dnu = 10.**(theta_asy.log_dnu)
+        
         d02 = self.get_d02(theta_asy)
+        
[…]
