+++ reggae/qtconsole.py
@@ -40,6 +40,8 @@ plt.rcParams['backend'] = 'QtAgg'
 class ReggaeDebugWindow(QtWidgets.QMainWindow):
 
     def __init__(self, input=None, session=None):
+        """Start the QT console.
+        """
 
         qapp = QtWidgets.QApplication.instance()
         self.qapp = qapp
@@ -63,6 +65,8 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
         qapp.exec()
 
     def init_ui(self):
+        """Initialize UI parameters
+        """
 
         self._main = QtWidgets.QSplitter()
         self.setCentralWidget(self._main)
@@ -432,6 +436,16 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
         self.print("Initialised Reggae Console.")
 
     def tqdm(self, gen, total=None, **kwargs):
+        """Make a custom tqdm
+
+        Parameters
+        ----------
+        gen: generator obj
+            Generator object to loop through
+        total: int, optional
+            Total number of iterations in the generator. Default is None.
+        """
+
         if total is None:
             total = len(gen)
         for i, _ in enumerate(gen):
@@ -440,9 +454,9 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
         self.progress.setValue(100)
 
     def sweep(self, i):
-        '''
+        """
         Parameter sweep over given bounds
-        '''
+        """
 
         field = FIELDS[i]
 
@@ -475,6 +489,8 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
         # self.threadpool.start(worker)
 
     def recompute(self):
+        """Make the recompute button
+        """
         self.sync_state()
         θreg, norm = self.get_state()
         self.replot(θreg, norm, label='Manual')
@@ -509,11 +525,10 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
                         d[_].remove()
                         del d[_]
 
-
     def replot(self, θreg, norm, label='Manual'):
-        '''
-        regenerate echelle diagrams
-        '''
+        """Make button to regenerate echelle diagrams
+        """
+
         ax = self._ax['echelle']
         ylims = ax.get_ylim()
         dnu = 10**self.reggae.theta_asy.log_dnu
@@ -610,8 +625,15 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
         ax.legend()
         ax.figure.canvas.draw()
 
-
     def load_pickle(self, description):
+        """Button to load pickle file
+        
+        Parameters
+        ----------
+        descripttion: str
+            Button tooltip.
+        """
+
         fname = QtWidgets.QFileDialog.getOpenFileName(self, description, ".", "Pickle files (*.pkl)")
         if fname[0] != '':
             with open(fname[0], 'rb') as f:
@@ -621,12 +643,28 @@ class ReggaeDebugWindow(QtWidgets.QMainWindow):
             return None
 
     def write_pickle(self, obj, description):
+        """Button to write pickle file
+        
+        Parameters
+        ----------
+        descripttion: str
+            Button tooltip.
+        """
+        
         fname = QtWidgets.QFileDialog.getSaveFileName(self, description, ".", "Pickle files (*.pkl)")
         if fname[0] != '':
             with open(fname[0], 'wb') as f:
                 dill.dump(obj, f)
 
     def load_pbjam(self, *, pbjam=None):
+        """Button to load pickled PBjam star class instance
+        
+        Parameters
+        ----------
+        pbjam: obj
+            PBjam.star.star class instance.
+        """
+        
         if pbjam is None:
[…]
