+++ docs/reggae.rst
@@ -1,0 +1,4 @@
+reggae
+^^^^^^
+.. automodule:: reggae.reggae
+    :members:
