+++ docs/api.rst
@@ -8,8 +8,9 @@ Reggae is built around
 .. The :mod:`~pbjam.session` and :mod:`~pbjam.star` modules are wrappers for many of the methods in :mod:`~pbjam.asy_peakbag` and :mod:`~pbjam.peakbag`, to help set up the analysis of one or more stars in a pipeline-like fashion.
 
 Modules in Reggae
-^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^
 .. toctree::
    :maxdepth: 0
 
    dipolestar
+   reggae
