+++ requirements.txt
@@ -1,7 +1,9 @@
 numpy >= 1.17.0
 matplotlib>=1.5.3
 lightkurve>=2.0.0
+pbjam
+astropy
 dynesty
 dill
 jax
-zeta
+zeta @ git+https://gitlab.com/darthoctopus/zeta@master
