### 2024-05-01 1344afe
right panel notes
