+++ docs/theta.rst
@@ -1,0 +1,4 @@
+theta
+^^^^^
+.. automodule:: reggae.theta
+    :members:
