### 2024-05-02 8dd21d7
docstrings for the classes themselves are required for their methods to show up automatically in the API documentation.
