+++ docs/qtconsole.rst
@@ -1,0 +1,4 @@
+qtconsole
+^^^^^^^^^
+.. automodule:: reggae.qtconsole
+    :members:
