### 2024-05-08 a5cb4b8
polish docstrings
