### 2024-05-08 4d71720
polish docstrings
