+++ reggae/dipolestar.py
@@ -36,6 +36,19 @@ class DipoleStar:
 
     @classmethod
     def from_pbjam(klass, star):
+        """ Import and prepare results from PBjam
+
+        Takes instances of the either the pbjam.star or pbjam.modeID.modeIDsampler 
+        classes and constructs a residual spectrum, by dividing out the l=2,0 (including
+        the background) or just the background respectively.
+
+        Parameters
+        ----------
+        klass: self
+            reggae.DipoleStar class self.
+        star: object
+            Instance of the pbjam.star or pbjam.modeID.modeIDsampler classes. 
+        """
 
         if isinstance(star, pbjam.star):
 
@@ -61,11 +74,15 @@ class DipoleStar:
         # theta_asy
 
         self.nu0, self.nu2, self.theta_asy = self._prepare_theta_asy(star)
+
         self.norders = len(self.nu0)
+        
         self.nmax = 10**(self.theta_asy.log_numax - self.theta_asy.log_dnu) - self.theta_asy.eps
+        
         self.l1model = PSDModel(self.f, self.norders)
 
         self.bounds = self.get_bounds()
+        
         self.l1model.n_g = self.select_n_g()
 
         self.soften = 1
@@ -109,67 +126,163 @@ class DipoleStar:
 
     @partial(jax.jit, static_argnums=(0,))
     def ptform(self, u):
-        '''
-        Turns coordinates on the unit cube into θ_reg object
-        '''
+        """ Turns coordinates on the unit cube into θ_reg object
+
+        Parameters
+        ----------
+        u: list
+            Point location in the unit hypercube.
+
+        Returns
+        -------
+        x: jnp.array
+            Point location in model parameter space.
+        """
+
         θ_reg = ThetaReg.prior_transform(u[:ThetaReg.dims], bounds=self.bounds)
+
         norm = jnp.array([self.bounds[-1][0] + u[-1] * (self.bounds[-1][1] - self.bounds[-1][0])])
+        
         return jnp.concatenate((θ_reg, norm))
 
     @partial(jax.jit, static_argnums=(0,))
-    def inv_ptform(self, θ_reg, norm):
-        '''
-        Turns θ_reg object into coordinates on the unit cube
-        '''
+    def inv_ptform(self, norm):
+        """  Turns θ_reg object into coordinates on the unit cube
+
+        Parameters
+        ----------
+        norm: jnp.array
+            Point location in model parameters space.
+        
+        Returns: jnp.array
+            Point location in the unit hypercube.
+        """
+
         u_reg = ThetaReg.inv_prior_transform(u[:ThetaReg.dims], bounds=self.bounds)
+        
         u_norm = jnp.array([(norm - self.bounds[-1][0]) / (self.bounds[-1][1] - self.bounds[-1][0])])
+        
         return jnp.concatenate((u_reg, u_norm))
 
     @partial(jax.jit, static_argnums=(0,))
-    def model(self, θ):
-        return self.l1model.l1model(self.theta_asy, ThetaReg(*θ[:9])) * θ[9] + 1
+    def model(self, theta):
+        """ Construct the model for the likelihood evaluation
+
+        Parameters
+        ----------
+        theta: jnp.array
+            Array of model parameters.
+
+        Returns
+        -------
+        mod: jnp.array
+            Spectrum model.
+        """
+
+        return self.l1model.l1model(self.theta_asy, ThetaReg(*theta[:9])) * theta[9] + 1
 
     @partial(jax.jit, static_argnums=(0,))
-    def ln_like(self, θ):
-        m = self.model(θ)
+    def ln_like(self, theta):
+        """ Evaluate the model log-likelihood at theta
+
+        The log-likelihood is the joint probability of observing
[…]
