### 2024-05-08 407afbf
polish docstrings
