+++ reggae/reggae/__init__.py
@@ -20,10 +20,10 @@ UNITS = {
 }
 
 class reggae():
-    '''
+    """
     jax'd numerical routines for generation of dipole mixed modes
-    and a PSD model.
-    '''
+    and a PSD model. 
+    """
 
     nu_to_omega = 2 * jnp.pi / 1e6
 
@@ -33,9 +33,27 @@ class reggae():
     @staticmethod
     @jax.jit
     def l1model(nu, nu1s, zeta, dnu, *, lw=None, amps=None):
-        '''
-        Generate a fake power spectrum from a set of modes (nu1s)
-        '''
+        """
+        Generate a power spectrum model from a set of modes (nu1s).
+
+        The model consists of a sequence of Lorentizan profiles.
+
+        Parmeters
+        ---------
+        nu: jnp.array
+            Frequency bins on which to compute the spectrum model.
+        nu1s: jnp.array
+            l=1 mode frequencies.
+        zeta: jnp.array
+            Mixing degree for the modes. 0 is completely p-like, 1 is 
+            completely g-like.
+        dnu: float
+            Large separation for the star.
+        lw: jnp.array, optional
+            Linewidths for the modes. Default is None. 
+        amps: jnp.array, optional
+            Mode amplitudes. Default is None.
+        """
 
         Hs = jnp.maximum(0, 1. - zeta)
         if amps is not None:
@@ -48,6 +66,7 @@ class reggae():
 
         for i in range(len(nu1s)):
             lorentzians += reggae._lor(nu, nu1s[i], Hs[i], modewidth1s[i])
+
         return lorentzians
 
     @staticmethod
@@ -118,12 +137,34 @@ class reggae():
     @staticmethod
     @jax.jit
     def asymptotic_nu_g(n_g, dPi0, max_N2, eps_g, alpha=0, numax=0., l=1):
-        '''
+        """
         Asymptotic relation for the g-mode frequencies
         in terms of a fundamental period offset (defined by the
         maximum Brunt-Vaisala frequency), the asymptotic g-mode period
         spacing, the g-mode phase offset, and an optional curvature term.
-        '''
+
+        Parameters
+        ----------
+        n_g: array-like
+            Array of radial orders at which the g-mode frequencies are computed
+        dPi0: float
+            The period spacing of the g-modes.
+        max_N2: float
+            The maximum of the Brunt-Vaisala frequency, used to define the fundamental
+            period offset.
+        eps_g: float 
+            Phase offset for the g-modes.
+        alpha: float, optional
+            Curvature factor for the g-mode periods. Default is 0.
+        numax: float, optional 
+            numax of the envelope. Default is 0
+        l=1: int, optional
+            Angular degree of the g-modes. Default is 1.
+
+        returns
+        -------
+    
+        """
 
         nu_to_omega = 2 * jnp.pi / 1e6
 
@@ -145,9 +186,23 @@ class reggae():
     @staticmethod
     @jax.jit
     def wrap_polyval2d(x, y, p):
-        '''
+        """
         Evaluate 2D polynomial, broadcasting shapes of x and y where appropriate.
-        '''
+
+        Parameters
+        ----------
+        x: array-like
+            x-coordinate for 2D polynomial.
+        y: array-like
+            y-coordinate for 2D polynomial.
+        p: array-like
+            Polynomial coefficients.
+            
+        Returns
+        -------
+        o: array-like
+            Result of evaluating the 2D polynomial.
+        """
 
         xx = (x + 0 * y)
 
[…]
