### 2024-05-08 e752c88
polish docstrings
