+++ docs/usage.md
@@ -5,25 +5,26 @@ Reggae is intended to be operated primarily through its GUI, although many GUI a
 - Executing the `reggae` module from the command line: `python -m reggae`. This will launch an empty instance of the GUI.
 - Instantiating a {class}`~reggae.qtconsole.ReggaeDebugWindow`. Passing it no arguments will launch an empty instance, while passing it a `pbjam.star.star` or {class}`~reggae.dipolestar.DipoleStar` object will open that object in the GUI.
 
-## Diagnostic Information
+## Left Panels: Diagnostic Information
 
 As a diagnostic utility, Reggae exposes a large amount of information to the viewer. This information is broken down in to several panels, which are switchable with the tab bar on the left side of the screen.
-
 :::{figure} images/ps.png
 :alt: Left Panel 1: Power Spectrum
 
 Left Panel 1: Power Spectrum
 :::
 
-The first panel shows the power spectrum. Like PBJam, Reggae uses the so-called SNR power spectrum (i.e. with a smooth background divided out) as its focus is on fitting individual peaks — this is shown in black. To further focus on the dipole modes, a model of the radial and quadrupole modes inherited from PBJam is further divided out — this is shown in blue. Thus, any residual peaks are those ignored by PBJam, which, a priori, are primarily dipole modes. However, the user should ensure that PBJam's model works as expected before proceeding further. The black and blue curves should agree except at the locations of PBJam's identified radial and quadrupole modes, at which locations there should be exactly two peaks at a time in the black curve that are not reflected in the blue.
+The first panel shows the power spectrum. Like `pbjam`, Reggae uses the so-called SNR power spectrum (i.e. with a smooth background divided out) as its focus is on fitting individual peaks — this is shown in black. To further focus on the dipole modes, a model of the radial and quadrupole modes inherited from `pbjam` is further divided out — this is shown in blue. Thus, any residual peaks are those ignored by `pbjam`, which, a priori, are primarily dipole modes. However, the user should ensure that `pbjam`'s model works as expected before proceeding further. The black and blue curves should agree except at the locations of `pbjam`'s identified radial and quadrupole modes, at which locations there should be exactly two peaks at a time in the black curve that are not reflected in the blue.
 
-Later steps in the fitting procedure produce PSD models as well. These are shown with various coloured curves.
+Later steps in the fitting procedure, using different optimisation schemes, may produce estimates of the parameters as well, each also corresponding to a PSD model. These are shown with various coloured curves.
 
 :::{figure} images/χ2.png
 :alt: Left Panel 2: $\chi^2$ Landscape
 
 Left Panel 2: $\chi^2$ Landscape
 :::
+
+This panel displays the values of $\chi^2 = -\log \mathcal{L}$ as a function of each of the optimisation parameters, holding the others constant at their current values, should the user initiate a parameter sweep (see Right Panel 2 below).
 
 :::{figure} images/echelle.png
 :alt: Left Panel 3: Echelle Diagrams
@@ -43,7 +44,9 @@ For the period-echelle power diagram in particular, the horizontal coordinate is
 Left Panel 4: Posterior Distributions
 :::
 
-## Tuning Parameters
+In addition to finding optimal point estimates, Reggae also allows the dipole-mode parameters to be determined by nested sampling. This panel simply displays the corner plot of the joint posterior distribution for the dipole-mode model parameters.
+
+## Right Panels: Tuning Parameters
 
[…]
