### 2024-05-08 3f367b6
adding installation instructions
