+++ joss/paper.md
@@ -28,13 +28,13 @@ csl-refs: yes
 
 # Summary
 
-`PBjam` [@{2021AJ....161...62N}] is a software instrument for fitting solar-like oscillation modes ("peakbagging\") in photometric power spectra returned from space missions like Kepler and TESS. Its upcoming second release [@pbjam2] supplements the simple model of the power spectrum used in the first version --- which included only radial and quadrupole ($\ell = 0, 2$) modes --- to additionally constrain more spectral features [e.g. @{2023A&A...676A.117N}]. Dipole ($\ell = 1$) modes, which had been specifically excluded in the initial version of the tool owing to their potential morphological complexity, are now specifically included. In keeping with the overall philosophy of `PBjam`'s design for $\ell = 0,2$, we are building a prior sample distribution of asymptotic parameters for these dipole modes. To assist in this task, we built a tool --- `Reggae`--- to manually fine-tune the dipole-mode model, and check the quality of both our initial guesses and fitted solutions.
+`PBjam` [@{2021AJ....161...62N}] is a software instrument for fitting solar-like oscillation modes ("peakbagging\") in photometric power spectra returned from space missions like Kepler and TESS. Its upcoming second release [@pbjam2] supplements the simple model of the power spectrum used in the first version --- which included only radial and quadrupole ($\ell = 0, 2$) modes --- to additionally constrain more spectral features [e.g. @{2023A&A...676A.117N}]. Dipole ($\ell = 1$) modes, which had been specifically excluded in the initial version of the tool owing to their potential morphological complexity, are now specifically included. Since the primary samples of the PLATO mission consist mainly of main-sequence and subgiant stars, `PBjam` implements a parameterisation of dipole modes suitable to these stars, outside the red-giant "asymptotic" regime. In keeping with the overall philosophy of `PBjam`'s design for $\ell = 0,2$, we will rely on an empirical prior sample of parameters describing these dipole modes to specify prior distributions. While the red giant regime has been extensively characterised observationally, the nonasymptotic construction here has not, requiring us to construct this prior sample ourselves. To assist in this task, we built a tool --- `Reggae`--- to manually fine-tune the dipole-mode model, and check the quality of both our initial guesses and fitted solutions.
 
 # Statement of Need
 
 An important part of this tuning is visual assessment of how well the data matches posterior samples for these parameters. Such asteroseismic visualisations often use the échelle power diagram near $\nu_{\mathrm{max}}$ as a diagnostic tool, with clearly-defined ridges emerging on this diagram for p-modes, such as in main-sequence stars. Gravitoacoustic mixed dipole modes in evolved stars, however, present more complicated features, making the distribution of mode power less visually intuitive in frequency space (see top frame of \autoref{fig:screenshot}). One may alternatively construct period-échelle power diagrams, correcting for mixed-mode coupling, to accommodate the asymptotic properties of g-modes, thereby again producing clear ridges. `Reggae` produces these visualisations from user-supplied trial values. This is useful for checking solutions of, e.g., the period spacing $\Delta\Pi_1$ --- inaccurate values result in slanted ridges, much like with inaccurate $\Delta\nu$ in traditional frequency échelle diagrams. Similarly, rotational splittings become easily identifiable, as are any perturbations due to magnetic fields.
 
[…]
