### 2024-05-14 darthoctopus/reggae#10 @sblunt
installation troubles
Hey folks! I'm one of the two JOSS reviewers for `reggae`. Sorry for the delay in getting started with this review! I hope I can help y'all out. I couldn't find your readthedocs page (but I see you're working on it), so for installation I followed your markdown file instructions.

I had trouble getting `reggae` installed in a fresh conda environment on my mac following the installation instructions, an issue which I've traced to the fact that the version of `pbjam` on pypi (1.0.0) is not up to date with the most recent github version (1.0.2), and `pbjam==1.0.0` fails to install. I was able to fix this by manually cloning and installing `pbjam==1.0.2` then retrying the `reggae` installation.

So, I see two possibilities:
1) encourage the `pbjam` folks to upload the most recent release to pypi
2) add this workaround of cloning and installing `pbjam` separately to your installation docs. If you go this route, I'd also suggest adding `pbjam>=1.0.2` to your requirements.txt.

One more suggestion on installation: it's generally recommended to use `pip install .` in a git cloned repo rather than `python setup.py install`. For developers, you can recommend `pip install -e .` if desired-- this flag will make it so that changes to the local copy of the code are automatically propagated to the local installation.

https://github.com/openjournals/joss-reviews/issues/6588
