### 2024-05-15 b9596fe
pip pyqt6 is apparently fine
