+++ docs/installation.md
@@ -1,6 +1,16 @@
 # Installation
+Before installing Reggae we strongly recommend you use a virtual environment. The simplest way to set one up the Python venv module:
+```
+python -m /path/to/new/virtual/environment
+``` 
+You will now need to activate the environment before you can install new Python packages. To do so just do:
+```
+source /path/to/new/virtual/environment/bin/activate
+```
+You can now freely install new packages without interfering with your previously installed packages.
+
 To install reggae do the following:
-- First make sure you have either PyQT 5 or 6 installed. Reggae relies on the Qt GUI console and so PyQt has to be installed. Preferably, this should be done using your system package manager or `conda`, rather than `pip`, since it links against compiled Qt libraries.
+ 
 - Next clone the `zeta` repository and install it by doing the following 
 ```
 cd path/to/my/repos
