+++ docs/usage.md
@@ -8,29 +8,24 @@ Reggae is intended to be operated primarily through its GUI, although many GUI a
 ## Left Panels: Diagnostic Information
 
 As a diagnostic utility, Reggae exposes a large amount of information to the viewer. This information is broken down in to several panels, which are switchable with the tab bar on the left side of the screen.
-:::{figure} images/ps.png
-:alt: Left Panel 1: Power Spectrum
 
-Left Panel 1: Power Spectrum
-:::
+### Power Spectrum
+
+![Left Panel 1: Power Spectrum](images/ps.png)
 
 The first panel shows the power spectrum. Like `pbjam`, Reggae uses the so-called SNR power spectrum (i.e. with a smooth background divided out) as its focus is on fitting individual peaks — this is shown in black. To further focus on the dipole modes, a model of the radial and quadrupole modes inherited from `pbjam` is further divided out — this is shown in blue. Thus, any residual peaks are those ignored by `pbjam`, which, a priori, are primarily dipole modes. However, the user should ensure that `pbjam`'s model works as expected before proceeding further. The black and blue curves should agree except at the locations of `pbjam`'s identified radial and quadrupole modes, at which locations there should be exactly two peaks at a time in the black curve that are not reflected in the blue.
 
 Later steps in the fitting procedure, using different optimisation schemes, may produce estimates of the parameters as well, each also corresponding to a PSD model. These are shown with various coloured curves.
 
-:::{figure} images/χ2.png
-:alt: Left Panel 2: $\chi^2$ Landscape
+### $\chi^2$ Landscape
 
-Left Panel 2: $\chi^2$ Landscape
-:::
+![Left Panel 2: $\chi^2$ Landscape](images/χ2.png)
 
 This panel displays the values of $\chi^2 = -\log \mathcal{L}$ as a function of each of the optimisation parameters, holding the others constant at their current values, should the user initiate a parameter sweep (see Right Panel 2 below).
 
-:::{figure} images/echelle.png
-:alt: Left Panel 3: Echelle Diagrams
+### Echelle Diagrams
 
-Left Panel 3: Echelle Diagrams
-:::
+![Left Panel 3: Echelle Diagrams](images/echelle.png)
 
 Given that the p- and g-modes are known to separately possess characteristic frequency and period spacings, respectively, it is conventional to examine the locations of these peaks in the power spectrum through the use of "echelle power diagrams". This third panel shows such diagrams, where the power spectrum is phase-folded with respect to these frequency or period spacings on the horizontal axis.
 
@@ -38,11 +33,9 @@ Reggae's mode identification step turns a small number of parameters — such as
 
 For the period-echelle power diagram in particular, the horizontal coordinate is further "stretched" through the use of the coordinate transform described in Mosser et al. 2012, 2015, 2017, 2018; Ong & Gehan 2023. In summary: in the asymptotic regime, mixed modes satisfy an eigenvalue equation parameterised by a single coupling strength $q$. If ansatz values of $q$, $\Delta\Pi$, and the pure p-mode frequencies are supplied, this allows the pure g-mode frequencies to be inferred from the mixed modes analytically. More generally, this specifies a coordinate transformation in the period coordinate mapping the mixed mode frequencies to the g-mode frequencies. As such, when applied to the period-echelle power diagram, it will map avoided crossings to vertical ridges, if the supplied values of $q$, $\Delta\Pi$, and the p-mode frequencies are accurate.
 
-:::{figure} images/posterior.png
-:alt: Left Panel 4: Posterior Distributions
+### Posterior Distribution
 
-Left Panel 4: Posterior Distributions
-:::
+![Left Panel 4: Posterior Distributions](images/posterior.png)
 
[…]
