### 2024-05-15 a6ffe03
adding test case to installation instructions
