### 2024-05-18 373ab9b
fix bug in DipoleStar.inv_ptform
