### 2024-05-18 a755aee
fix docstring for ThetaReg.inv_ptform
