### 2024-05-18 c29dd31
various fixes for changes to function signatures???? when did these happen
