### 2024-05-18 darthoctopus/reggae#9 @darthoctopus
Hi Sylvain,

Thanks for the notes. We have taken the following steps to address your feedback:

- We have added an excerpt of the paper's statement of need to the README.md file.
- We have added a section to the README.md file with quick installation instructions, and also have created a full set of documentation files with more detailed instructions if needed.
- We have also added a more comprehensive description of the GUI to our documentation folder.
- We have opted to use Sphinx to generate an API reference automatically. This is automatically generated on readthedocs from the docstrings of our functions and classes, which we have now also populated.
- We have updated our testing data to overcome a pandas incompatibility as part of fixing the packaging issues that Sarah has pointed out, and now include instructions for running reggae on it, both in our documentation, and as a README file in the test folder.
- We have added a set of community contribution guidelines (in particular a contributor code of conduct).

Please let us know what other changes you believe we ought to make for compliance with these guidelines.
