### 2024-05-18 darthoctopus/reggae#10 @darthoctopus
Hi Sarah,

We've restructured our requirements.txt to explicitly require pbjam 1.0.3, which is now on PyPI. We have also rewritten our installation instructions to include your suggestion to use pip rather than setup.py.

I have tested that the new instructions work on a fresh python3.10 environment --- unfortunately, pbjam currently has a dependency on a rather old version of scipy, and reggae won't be installable on newer versions of python until pbjam 2 is released. Please try again and let us know if you encounter any issues.
