### 2024-05-21 darthoctopus/reggae#10 @sblunt
Sounds great! I'd recommend adding a note that you require python<=3.10 to the install page as well. But otherwise the updated install page looks good to me.
