### 2024-05-21 darthoctopus/reggae#10 @sblunt
One more thing-- since jax is one of your dependencies, I think it's worth noting that some of your users may have trouble installing depending on their OS (see [here](https://jax.readthedocs.io/en/latest/installation.html#supported-platforms)), and you could potentially include a link to the jax install page under an "common installation problems FAQ" or soemthing like that
