### 2024-05-22 darthoctopus/reggae#10 @darthoctopus
Hi Sarah,

We've added both a note about python versions, and a section for FAQs with a reference to the jax OS support page, to the installation section of our documentation. Thanks for these suggestions.
