+++ docs/installation.md
@@ -1,6 +1,6 @@
 # Installation
 
-**We presently only officially support installating Reggae under Python 3.10 and below**. Before installing Reggae we strongly recommend you use a virtual environment. The simplest way to set one up using the Python `venv` module:
+**We presently only officially support installating Reggae under Python 3.10 and below**. Before installing Reggae we strongly recommend you use a virtual environment. The simplest way to set one up is using the Python `venv` module:
 ```
 python -m /path/to/new/virtual/environment
 ``` 
@@ -41,4 +41,4 @@ This should open up an UI featuring the example data.
 
 - What do I need to install in order to accelerate these calculations on my GPU?
 
-Please check that your OS is on `jax`'s supported list of systems and hardware configurations, which can be found [here](https://jax.readthedocs.io/en/latest/installation.html#supported-platforms). In summary, while performing `jax` calculations on the CPU is supported on all platforms, GPU acceleration is only available on limited combinations of hardware and operating systems.
+Please check that your OS is on `jax`'s supported list of systems and hardware configurations, which can be found [here](https://jax.readthedocs.io/en/latest/installation.html#supported-platforms). In summary, while performing `jax` calculations on the CPU is supported on all platforms, GPU acceleration is only available on limited combinations of hardware and operating systems. Our `requirements.txt` file assumes only CPU support via `jaxlib`; Windows users may need additional external requirements for it to work.
