### 2024-05-22 darthoctopus/reggae#9 @sybreton
Hi Joel,

Many thanks !
I have started reviewing the additions you made in order to validate the items of the checklist.
Concerning the module installation, it is working fine on my side, I just have a small suggestion: as `reggae` requires `zeta`, it could be good to include it directly in the `requirements.txt` by pointing to the GitLab repo, the following line should work:
`zeta @ git+https://gitlab.com/darthoctopus/zeta.git`
Otherwise you should state on the `README.md` that `zeta` is required (a user that did not follow the link to the detailed installation instruction might likely omit this installation step otherwise).
