### 2024-05-22 fba11ff
Harmonise module name style and add link to cited papers and repo
