### 2024-05-22 darthoctopus/reggae#9 @sybreton
Concerning the statement of need, I created the following pull request in order to include links towards the cited papers and the PBjam repo (I also harmonised the way module names are written), feel free to merge it !
https://github.com/darthoctopus/reggae/pull/11
