+++ docs/installation.md
@@ -10,16 +10,7 @@ source /path/to/new/virtual/environment/bin/activate
 ```
 You can now freely install new packages without interfering with your previously installed packages.
 
-To install reggae do the following:
- 
-- Next clone the `zeta` repository and install it by doing the following 
-```
-cd path/to/my/repos
-git clone https://gitlab.com/darthoctopus/zeta.git
-cd zeta
-pip install -e .
-```
-- Now clone the `Reggae` repository and install it by doing the following
+To install `reggae`, clone this repository and install it by doing the following:
 ```
 cd path/to/my/repos
 git clone https://github.com/darthoctopus/reggae.git
@@ -42,3 +33,14 @@ This should open up an UI featuring the example data.
 - What do I need to install in order to accelerate these calculations on my GPU?
 
 Please check that your OS is on `jax`'s supported list of systems and hardware configurations, which can be found [here](https://jax.readthedocs.io/en/latest/installation.html#supported-platforms). In summary, while performing `jax` calculations on the CPU is supported on all platforms, GPU acceleration is only available on limited combinations of hardware and operating systems. Our `requirements.txt` file assumes only CPU support via `jaxlib`; Windows users may need additional external requirements for it to work.
+
+- I am getting a package conflict for a module named `zeta` when attempting to install.
+
+`reggae` relies on a package named `zeta` for asymptotic mode-coupling calculations and generation of stretched period-echelle power diagrams. Unfortunately, that name is already taken on PyPI, so it is not available there and has to be installed from its gitlab repository. If installation via `requirements.txt` does not work, it may also be installed manually with the following commands:
+
+```
+cd path/to/my/repos
+git clone https://gitlab.com/darthoctopus/zeta.git
+cd zeta
+pip install -e .
+```
