### 2024-05-22 darthoctopus/reggae#9 @darthoctopus
Thanks for this feedback! We have made your suggested changes to the `requirements.txt` and tested that they work, and added some notes to the installation FAQs. We have also merged your pull request — thanks for taking the time to do that.
