### 2024-05-22 90d3771
add links to other papers in readme
