+++ README.md
@@ -32,7 +32,7 @@ For more detailed usage instructions, please refer to our [documentation](https:
 
 ## Generative model for mode frequencies
 
-We implement a generative model for dipole gravitoacoustic mixed modes using the parameterisation of Ong & Basu (2020). At present, the frequency-dependent coupling strength is described with two parameters (one for each of the two matrices entering into the parameterisation), with a conversion to the asymptotic $q$ provided by an expression in Ong & Gehan (2023). This expression is in turn used to generate stretched echelle power plots for diagnostic purposes.
+We implement a generative model for dipole gravitoacoustic mixed modes using the parameterisation of [Ong & Basu (2020)](https://ui.adsabs.harvard.edu/abs/2020ApJ...898..127O/abstract). At present, the frequency-dependent coupling strength is described with two parameters (one for each of the two matrices entering into the parameterisation), with a conversion to the asymptotic $q$ provided by an expression in [Ong & Gehan (2023)](https://ui.adsabs.harvard.edu/abs/2023ApJ...946...92O/abstract). This expression is in turn used to generate stretched echelle power plots for diagnostic purposes.
 
 In full, the generative model accepts the following parameters:
 
