### 2024-05-23 f8142b2
let's not rely on lightkurve for echelle diagrams
