### 2024-05-23 ab742ad
save checkbox states to session pickle too
