### 2024-05-23 fbdf7e7
set tight layout so that axis labels don't get cut off on smaller screens
