### 2024-05-27 8f40862
\mod doesn't work with usetex=False
