### 2024-05-28 darthoctopus/reggae#9 @sybreton
Dear Joel,

Thanks again for the significant additions you made for the documentation. I think a few elements are still needed to tick the remaining boxes of the "Documentation" section:
- "Example usage": testing the code, I was a little bit lost concerning what I should do after I run `python -m reggae`. The GUI is working fine, but I hade no idea how I can e.g. import my own data to exploit the module functionality. Could you add in your documentation a step-by-step tutorial detailing a typical analysis case, explaining how to import the required inputs (any fits file with a frequency vector and corresponding PSD ? some PBJam outputs ?) and presenting an example of parameter-tuning with the options provided in the GUI ? (I would suggest to illustrate this with the case you included in `test.py`)
- "Functionality documentation": alternatively to the GUI, you mention two classes, `ReggaeDebugWindow` and `DipoleStars`, could you use the docstrings you wrote for these classes and related function to generated an additional API reference page in your documentation.
- "Community guidelines": in addition to the Code of conduct you added, could you add a very short section in your documentation stating that e.g. 1) people who might want to contribute should fork the repo and open pull request (additional suggestion about this but it is totally ok for me if you decide not to go for it: you can setup an automated CI/CD pipeline so people who modified the code have a normalised way to check that they did not broke anything when they make the pull request), 2)3) people who want to report problems with the code or seek support should open issue here on this GitHub repo.
