### 2024-06-11 #6588 @sybreton
@dfm - Sorry I have been wanting to answer you for a few day already ! No particular issues to mention on my side, once @darthoctopus will have added the few items I required for the documentation (https://github.com/darthoctopus/reggae/issues/9#issuecomment-2135446923) I should be able to tick the missing boxes for this part. I started reading the software paper and it is already in a very good state, so things should go smoothly on this side too.
