### 2024-06-11 darthoctopus/reggae#10 @sblunt
Ok, I was finally able to get the code working by downgrading my versions of `PyQt6-Qt6` and `PyQt6` to 6.4.0! I was getting a segfault using `PyQt6-Qt6`>=6.5 and `PyQt6`>=6.5. I'm not sure if I can make an actionable suggestion, other than it might be a good thing to add a suggestion along those lines to the installation FAQs?
