### 2024-06-11 darthoctopus/reggae#10 @sblunt
Oh, one more thing-- at one point in my debugging process, I tried running the code in Python 3.9 and ran into a different error. If you're also seeing that, then I'd suggest adding to the docs that you support Python 3.10 only.
