### 2024-06-11 darthoctopus/reggae#12 @sblunt
feedback on JOSS paper text
Hey folks! I've given your JOSS text a close read, and I have a few suggestions based on the checkbox items of the JOSS review process. I'm a total astroseismology noob, so please forgive my ignorance!

Summary:

1. It was a little unclear to me upon first few reads where `PBJam` and `reggae` overlap (and don't) in functionality. I think that can be clarified a bit in the summary and statement of need. Is it correct to state that `PBJam` performs full fits, while `reggae` visualizes them and performs small fits to reduce prior search volume?

2. "We are building a prior sample distribution of asymptotic parameters for these dipole modes" -- as an outsider to the field, I have no idea what this means. Are you constraining the population of asymptotic parameters for dipole modes across all stars, one star, a subset of stars? How do you expect knowledge will be used by the community?

3.  In the first paragraph of the statement of need, I think it would be helpful to add a reference the reader can refer to for more info/examples about period-echelle power diagrams and why they are more helpful than traditional echelle diagrams for mixed-mode oscillators. I felt that I had to do a lot of digging to get the background info I needed to understand this section. Also, the way you correct for mixed-mode coupling is described in the next section, right? It may help to just include a parenthetical "see next section" when you introduce the correction.

4. "reggae picks up where pbjam leaves off" -- this sentence confused me. Maybe something like "reggae visualizes the residuals to an l=0,2 model computed by PBJam, and can either be used to visualize or manually tune an initial guess for the l=1 mode"

5. I don't think you discussed "state of the field" anywhere in the JOSS text. Could you add a short section summarizing similar software, or saying something like "we're not aware of other software for visualizing dipole mode fits."

A few other documentation suggestions:

1. In your code of conduct, please add an email address where it says [INSERT EMAIL ADDRESS]

2. I agree with @sybreton's most recent comment about expanding the sample usage. In particular, I think it would be helpful to walk the user through generating the figure you show in the JOSS paper.

3. Have you gotten your readthedocs page working? I see that you've got sphinx documentation, and I can see reggae on readthedocs when I google it, but when I click the link, I get a 404 error. Once you get the docs up and running, I'd suggest linking to them from your readme page too.

https://github.com/openjournals/joss-reviews/issues/6588
