### 2024-06-25 f1b1ef2
__match_args__ was only introduced in 3.10
