### 2024-06-25 #6588 @darthoctopus
Hi all,

Apologies for the delay. Working my way through these issues now — I should have responses ready to the currently open issues by the end of the week.
