### 2024-06-25 darthoctopus/reggae#10 @darthoctopus
Hi Sarah,

Thanks for these notes.

- I'm able to reproduce your issue on Python 3.9 and have pushed a fix.
- I am unfortunately *not* able to reproduce your Qt issue — e.g. on both Python 3.10 and 3.9, I have been able to get our code to work on PyQt6 6.7.0. Could you please provide more information about your environment? This would be helpful for writing the entry in the FAQ that you suggest.
