### 2024-06-26 60fc18b
fix some docstring formatting
