### 2024-06-26 34fe4e1
filename fix for theta and chi2 pngs
