### 2024-06-26 3bcb744
switch to using autoapi rather than autodocs
