+++ docs/conf.py
@@ -15,7 +15,7 @@ release = '0.0.1'
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    'sphinx.ext.autodoc',
+    'autoapi.extension',
     'sphinx.ext.mathjax',
     'myst_parser'
 ]
@@ -27,6 +27,8 @@ myst_enable_extensions = [
     "smartquotes",
     "strikethrough",
 ]
+
+autoapi_dirs = ['../reggae']
 
 templates_path = ['_templates']
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
