+++ docs/theta.rst
@@ -1,4 +1,0 @@
-theta
-^^^^^
-.. automodule:: reggae.theta
-    :members:
