+++ docs/dipolestar.rst
@@ -1,4 +1,0 @@
-dipolestar
-^^^^^^^^^^
-.. automodule:: reggae.dipolestar
-    :members:
