+++ docs/reggae.rst
@@ -1,4 +1,0 @@
-reggae
-^^^^^^
-.. automodule:: reggae.reggae
-    :members:
