+++ .readthedocs.yaml
@@ -30,5 +30,3 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
-   - method: pip
-     path: .
