### 2024-06-26 9d6a950
readthedocs badge & links
