### 2024-06-26 09b5633
implement sylvain's suggestions for the readme
