### 2024-06-26 c14a5b7
include parts of the readme in the main page of the readthedocs
