### 2024-06-26 c1afb97
update test to automatically load saved GUI session (and fix kwarg bug)
