### 2024-06-27 darthoctopus/reggae#10 @darthoctopus
Hi Sarah,

Thanks for the description and the error dump. This seems like an issue we've run into previously on some linux systems where only Wayland is installed by default for user GUI sessions. This is unfortunately not something that we can reasonably anticipate, as it requires intervention through the package manager for some compiled dependencies to work (or, as you have found, using alternative versions for certain packages). I will add a note about this to the FAQs.
