### 2024-06-27 b524f37
update FAQ with Qt thing
