### 2024-06-27 darthoctopus/reggae#10 @darthoctopus
Hi Sarah,

I've added an FAQ entry. Thanks for the heads up about your installation troubles!
