### 2024-06-27 darthoctopus/reggae#9 @darthoctopus
Hi Sylvain,

Thanks for your continued constructive input. We've added some additional documentation that might be helpful:

- We now automatically load the included saved session for the test case (this was not previously done), which we hope will be a little more user-friendly: this means that the GUI for the test case will be initially populated with a reasonable initial guess for the star, rather than the blanket defaults.
- We have written a hopefully helpful Quick Start guide describing how to go from PBJam peakbagging results to initial guesses of the dipole-mode parameters and, eventually, saved DipoleStar objects.
- We have finally managed to get our readthedocs up and running. Part of this delay was because `reggae.readthedocs.io` was already taken by an unrelated project, so we have instead opted to host our documentation at <https://pb-reggae.readthedocs.io/>. This also contains automatically-generated API documentation.
- Finally, we have updated our description of the community guidelines with your suggested changes
