### 2024-06-27 darthoctopus/reggae#12 @darthoctopus
Hi Sarah,

Thanks for the constructive suggestions, and much-needed outside perspective! I will work on addressing your major points about the paper per se later this evening, but for now, I think we have managed to address your minor points. You may have seen some of these in our response to @sybreton, but, for completeness:
- I have added both my and Martin's email addresses to the Code of Conduct
- We now have a (hopefully sufficiently detailed) Quick Start guide containing a rough outline of the steps we followed to produce fits to the power spectrum like the one provided in the test case
- We've managed to get our documentation hosted at <https://pb-reggae.readthedocs.io/>, since the `reggae` short name was already taken by an unrelated project.
