### 2024-06-27 4e7d825
tweaks to readme files
