+++ joss/paper.md
@@ -28,19 +28,19 @@ csl-refs: yes
 
 # Summary
 
-`PBjam` [@{2021AJ....161...62N}] is a software instrument for fitting solar-like oscillation modes ("peakbagging\") in photometric power spectra returned from space missions like Kepler and TESS. Its upcoming second release [@pbjam2] supplements the simple model of the power spectrum used in the first version --- which included only radial and quadrupole ($\ell = 0, 2$) modes --- to additionally constrain more spectral features [e.g. @{2023A&A...676A.117N}]. Dipole ($\ell = 1$) modes, which had been specifically excluded in the initial version of the tool owing to their potential morphological complexity, are now specifically included. Since the primary samples of the PLATO mission consist mainly of main-sequence and subgiant stars, `PBjam` implements a parameterisation of dipole modes suitable to these stars, outside the red-giant "asymptotic" regime. In keeping with the overall philosophy of `PBjam`'s design for $\ell = 0,2$, we will rely on an empirical prior sample of parameters describing these dipole modes to specify prior distributions. While the red giant regime has been extensively characterised observationally, the nonasymptotic construction here has not, requiring us to construct this prior sample ourselves. To assist in this task, we built a tool --- `Reggae`--- to manually fine-tune the dipole-mode model, and check the quality of both our initial guesses and fitted solutions.
+`PBjam` [@{2021AJ....161...62N}] is a software instrument for fitting solar-like oscillation modes ("peakbagging\") in power spectra from space-based photometry. Its upcoming second release [@pbjam2] supplements the simple power-spectrum model used in the first version --- which included only radial and quadrupole ($\ell = 0, 2$) modes --- to additionally constrain other features [e.g. @{2023A&A...676A.117N}]. Dipole ($\ell = 1$) modes, which had been specifically excluded in the initial version of the tool owing to their complexity, are now specifically included. Since the primary samples of the PLATO mission consist mainly of main-sequence and subgiant stars [@plato], `PBjam` implements a parameterisation of dipole mixed-mode frequencies --- as described by their overtone spacings, boundary conditions, and other stellar properties --- that is suitable to these stars, outside the red-giant "asymptotic" regime. In keeping with the overall philosophy of `PBjam`'s design for $\ell = 0,2$, `PBjam` 2 will specify prior distributions on these parameters through an empirical prior sample. While the red giant regime has been extensively characterised observationally, the nonasymptotic construction here has not, requiring us to construct this prior sample ourselves. To assist in this task, we built a tool --- `Reggae`--- to manually fine-tune and fit the dipole-mode model, and check the quality of both our initial guesses and fitted solutions.
 
 # Statement of Need
 
[…]
