+++ joss/biblio.bib
@@ -194,6 +194,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{ong_rotation_2022,
+       author = {{Ong}, J.~M. Joel and {Bugnet}, Lisa and {Basu}, Sarbani},
+        title = "{Mode Mixing and Rotational Splittings. I. Near-degeneracy Effects Revisited}",
+      journal = {\apj},
+     keywords = {Asteroseismology, Computational methods, Theoretical techniques, Stellar rotation, Stellar magnetic fields, 73, 1965, 2093, 1629, 1610, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2022,
+        month = nov,
+       volume = {940},
+       number = {1},
+          eid = {18},
+        pages = {18},
+          doi = {10.3847/1538-4357/ac97e7},
+archivePrefix = {arXiv},
+       eprint = {2210.01928},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJ...940...18O},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{ong_rotation_2023,
     author = {{Ong}, J.~M. Joel and Charlotte Gehan},
     title = {Mode Mixing and Rotational Splittings. II. Reconciling Different Approaches to Mode Coupling},
@@ -267,3 +286,21 @@ archivePrefix = {arXiv},
     title = {PBJam 2.0: Mixed Modes are Everywhere, But We've Got It Sorted},
     year = {in prep.}
 }
+
+@ARTICLE{plato,
[…]
