+++ joss/biblio.bib
@@ -304,3 +304,48 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{gehan_core_2018,
+       author = {{Gehan}, C. and {Mosser}, B. and {Michel}, E. and {Samadi}, R. and {Kallinger}, T.},
+        title = "{Core rotation braking on the red giant branch for various mass ranges}",
+      journal = {\aap},
+     keywords = {asteroseismology, methods: data analysis, stars: interiors, stars: oscillations, stars: rotation, stars: solar-type, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2018,
+        month = aug,
+       volume = {616},
+          eid = {A24},
+        pages = {A24},
+          doi = {10.1051/0004-6361/201832822},
+archivePrefix = {arXiv},
+       eprint = {1802.04558},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018A&A...616A..24G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@BOOK{aerts_asteroseismology_2010,
+       author = {{Aerts}, Conny and {Christensen-Dalsgaard}, J{\o}rgen and {Kurtz}, Donald W.},
+        title = "{Asteroseismology}",
+         year = 2010,
+          doi = {10.1007/978-1-4020-5803-5},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010aste.book.....A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{mosser_probing_2012,
+       author = {{Mosser}, B. and {Goupil}, M.~J. and {Belkacem}, K. and {Michel}, E. and {Stello}, D. and {Marques}, J.~P. and {Elsworth}, Y. and {Barban}, C. and {Beck}, P.~G. and {Bedding}, T.~R. and {De Ridder}, J. and {Garc{\'\i}a}, R.~A. and {Hekker}, S. and {Kallinger}, T. and {Samadi}, R. and {Stumpe}, M.~C. and {Barclay}, T. and {Burke}, C.~J.},
+        title = "{Probing the core structure and evolution of red giants using gravity-dominated mixed modes observed with Kepler}",
+      journal = {\aap},
+     keywords = {stars: oscillations, stars: interiors, stars: evolution, stars: late-type, methods: data analysis, asteroseismology, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2012,
+        month = apr,
+       volume = {540},
+          eid = {A143},
+        pages = {A143},
+          doi = {10.1051/0004-6361/201118519},
+archivePrefix = {arXiv},
+       eprint = {1203.0689},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012A&A...540A.143M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
