+++ joss/paper.md
@@ -34,7 +34,7 @@ csl-refs: yes
 
 Before mode frequencies may be extracted from the power spectrum, specific peaks in it must be identified as dipole modes. An important part of this identification is visual assessment of how well the predicted mode frequencies correspond to actually observed peaks. `Reggae` produces these visualisations from user-supplied trial values. This is useful for checking solutions of, e.g., the period spacing $\Delta\Pi_1$ --- inaccurate values result in slanted ridges on period-echelle diagrams [e.g. @mosser_probing_2012], much like with inaccurate $\Delta\nu$ in traditional frequency-échelle diagrams [e.g. @aerts_asteroseismology_2010]. Similarly, rotational splittings become easily identifiable [e.g. @gehan_core_2018], as are any perturbations due to magnetic fields [@hatt].
 
-Since these global parameters must be supplied for dipole-mode identification, we have constrained them for a preliminary sample of subgiants [@pbjam2], and also for a large sample of low-luminosity red giants [@hatt]. We found `Reggae` very helpful both for these tuning and visualisation tasks, and also as a didactic aid to understanding the dipole mixed-mode parameters. Moreover, no other tools currently exist for performing this task in the nonasymptotic parameterisation that `PBjam` will use. As such, we release `Reggae` publicly in advance of the second `PBjam` version, as we believe the community will benefit from access to such a visualisation tool. This will also assist future users of `PBjam` in devising ad-hoc prior constraints on the mixed-mode parameters, should they wish to perform mode identification for anomalous stars.
+Since these global parameters must be supplied for dipole-mode identification, we have constrained them for a preliminary sample of subgiants [@pbjam2], and also for a large sample of low-luminosity red giants [@hatt]. We found `Reggae` very helpful both for these tuning and visualisation tasks, and also as a didactic aid to understanding the dipole mixed-mode parameters. Moreover, no other tools currently exist for performing these tasks in the nonasymptotic parameterisation that `PBjam` will use. As such, we release `Reggae` publicly in advance of the second `PBjam` version, as we believe the community will benefit from access to such a visualisation tool. This will also assist future users of `PBjam` in devising ad-hoc prior constraints on the mixed-mode parameters, should they wish to perform mode identification for anomalous stars.
 
 # Modeling the Oscillation Spectrum
 
