### 2024-06-27 darthoctopus/reggae#12 @darthoctopus
Hi Sarah,

I have updated the JOSS paper (in the paper-specific branch) to hopefully address your comments. In particular:

1. We now (hopefully) make clear that, whereas pbjam is intended for fitting individual mode frequencies, reggae is intended instead to assist in visual assessment of, and deriving preliminary constraints on, the overall properties of the dipole modes (e.g. characteristic spacings, rotation rates, phase offsets instead of individual frequencies).
2. The phrasing for this sentence is now "`PBjam` 2 will specify prior distributions on these parameters empirically, through predetermined values found for existing samples of solar-like oscillators." --- this better reflects how this population of parameters, as determined from primarily Kepler and TESS, will be applied to PLATO data.
3. Since this level of technical detail doesn't seem to be necessary in a statement of need, and (as you observe) is already provided in the subsequent sections, we have simplified this passage significantly, and added references to each application described here.
4. This sentence has been revised to hopefully be more immediately descriptive of what reggae actually does.
5. We've added some words to this effect to the statement of need: "no other tools currently exist for performing this task in the nonasymptotic parameterisation that PBjam will use"

The revised manuscript is unfortunately a little longer than the 1000 word limit, however.
