### 2024-07-04 darthoctopus/reggae#9 @sybreton
Hi Joel,

Thanks a lot for the addition !
Just one final suggestion that I think will be useful for the users: in the quickstart tutorial, at this stage: https://pb-reggae.readthedocs.io/en/latest/quickstart.html#creating-the-dipolestar-object, could you add an example line of code corresponding to each use case ? Which would be `python -m reggae` in the first case, and the initialisation of the DipoleStar object in the other. I think it would be good at this stage to explicitly point towards the example data you included in the `test` directory so the users can make their own tests with the GUI using this dataset.
