### 2024-07-04 darthoctopus/reggae#12 @sybreton
Hi Sarah and Joel,

Concerning the paper, I just have a few minor comments so I think it is better I send them here rather than opening another issue:

"_PBjam [@{2021AJ....161...62N}] is a software instrument for fitting solar-like oscillation modes ("peakbagging")_"
--> concerning solar-like oscillations maybe add a reference towards a recent general review of the field, e.g. García & Ballot 2019 (https://ui.adsabs.harvard.edu/abs/2019LRSP...16....4G/abstract)

"_Since the primary samples of the PLATO mission consist mainly of main-sequence and subgiant stars [@plato], PBjam implements a parameterisation of dipole mixed-mode frequencies --- as described by their overtone spacings, boundary conditions, and other stellar properties --- that is suitable to these stars_"
--> it is ambiguous that mixed modes are expected only for subgiant and not for main-sequence stars, I think this sentence should be slightly rephrased to avoid any confusion on this aspect.

"_the nonasymptotic construction_"
--> for clarity towards non-specialists, it would be good to recall at this point that the non-asymptotic regime corresponds to subgiant stars.

How far are you currently from the 1000 words limit ?
