+++ joss/biblio.bib
@@ -349,3 +349,22 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{garcia_asteroseismology_2019,
+       author = {{Garc{\'\i}a}, Rafael A. and {Ballot}, J{\'e}r{\^o}me},
+        title = "{Asteroseismology of solar-type stars}",
+      journal = {Living Reviews in Solar Physics},
+     keywords = {Asteroseismology, Stellar oscillations, Solar analogs, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = sep,
+       volume = {16},
+       number = {1},
+          eid = {4},
+        pages = {4},
+          doi = {10.1007/s41116-019-0020-1},
+archivePrefix = {arXiv},
+       eprint = {1906.12262},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019LRSP...16....4G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
