### 2024-07-05 darthoctopus/reggae#12 @darthoctopus
Hi Sylvain,

Thanks for these notes. I've made the following changes to the draft:

1. The first sentence now reads "`PBjam` is a software instrument for fitting normal modes ("peakbagging\") in power spectra from space-based photometry of solar-like oscillators [e.g. @garcia_asteroseismology_2019]" , incorporating your suggested reference.
2. The phrasing of this sentence has been changed from "suitable to these stars" to "reduces to pure p-modes in the former, and is suitable to the latter".
3. This sentence is now rephrased to read "the nonasymptotic construction for subgiants here".

Presently the manuscript (including author metadata but not references) is 1080 words long.
