### 2024-07-10 #6588 @sybreton
On my side I validated all the items of the checklist and I am also completely satisfied with the additions that have been made!
