+++ joss/biblio.bib
@@ -276,9 +276,22 @@ archivePrefix = {arXiv},
 }
 
 @article{lindsay,
-    author = {Christopher J. Lindsay and J. M. Joel Ong and Sarbani Basu},
-    title = {Fossil Signatures of Main-sequence Convective Core Overshoot in Subgiant Stars Estimated through Asteroseismic Analyses},
-    year = {submitted to ApJ}
+       author = {Christopher J. Lindsay and J. M. Joel Ong and Sarbani Basu},
+        title = "{Fossil Signatures of Main-sequence Convective Core Overshoot Estimated through Asteroseismic Analyses}",
+      journal = {ApJ},
+     keywords = {Asteroseismology, Stellar physics, Stellar evolution, Stellar convective zones, Subgiant stars, 73, 1621, 1599, 301, 1646, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2024,
+        month = apr,
+       volume = {965},
+       number = {2},
+          eid = {171},
+        pages = {171},
+          doi = {10.3847/1538-4357/ad2ae5},
+archivePrefix = {arXiv},
+       eprint = {2402.12461},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJ...965..171L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{pbjam2,
