### 2024-07-12 9e27c2f
zenodo and JOSS badges
