### 2024-07-13 #6588 @darthoctopus
Hi Dan,

I've had a look at the merge request and it looks reasonable — the backslash before the closing quote was needed to get inara to compile a pdf on my work machine, but it compiles now on my laptop without it. Thanks for helping us with the references — I must confess to not having noticed them missing before. I have merged it now.

I have incremented the version number from 0.0.1 to 0.1.0, and deposited the repository on Zenodo; the DOI is 10.5281/zenodo.12730547 .
