### 2024-07-13 #6588 @dfm
@darthoctopus — Thanks. One last edit that I missed: can you add the `# References` section heading back to the bottom of the manuscript?

Also: please update the Zenodo metadata to have the same title as the manuscript.
