### 2024-07-13 #6588 @darthoctopus
Hi Dan,

Sure, I have just now made both of these changes (sorry for the delay)
