### 2024-07-13 #6588 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
